Fixes for BUG: If I remove an item, and then submit the sale, it presents the static page, but doesn't ask me to confirm or re-edit the details.  When I refresh the page, it tells me the sale is in error, which obviously means some records were removed and others weren't.   It looks like the same bug mentioned in section 5.a (the sale_item_service_landline_business record is deleted).  So this problem will be rectified when 5.a is fixed, apart from the fact that you still don't get the "Confirm" and "Edit" buttons
Also, fixed another bug associated with the deletion of contact methods

+++ html/admin/classes/json/handler/sale/JSON_Handler_Sale.php
@@ -304,9 +304,12 @@ class JSON_Handler_Sale extends JSON_Handler
 					throw new Exception("No preferred contact method selected.");
 				}
 
-				foreach ($arrContactMethods as $unwantedContactMethod)
+				if (!$bolValidateOnly)
 				{
-					$unwantedContactMethod->delete();
+					foreach ($arrContactMethods as $unwantedContactMethod)
+					{
+						$unwantedContactMethod->delete();
+					}
 				}
 			}
 			
@@ -392,11 +395,11 @@ class JSON_Handler_Sale extends JSON_Handler
 				}
 			}
 			
-			if ($bolValidateOnly)
+			if (!$bolValidateOnly)
 			{
 				foreach($arrItems as $removedItem)
 				{
-					$removedItem->cancel();
+					$removedItem->cancel($dealer->id);
 				}
 			}
